chore: no report

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,5 @@
 #include "function_pointers.h"
+#include <stdlib.h>
 
 /**
  * array_iterator - executes a function as a parameter on each element of an
@@ -14,6 +15,6 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 	if (array == NULL || size < 0 || action == NULL)
 		return;
 
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 		action(array[i]);
 }
